+++ Toggly.FeatureManagement.Flutter/toggly/lib/src/services/secure_storage_service.dart
@@ -4,6 +4,7 @@ enum SecureStorageKeys {
   featureFlagsCache,
   deviceId,
   jwks,
+  etag,
 }
 
 /// Data representation to be stored in/retrieved from cache.